38 fail test

+++ ft_select_arg.c
@@ -6,7 +6,7 @@
 /*   By: rhunders <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/25 00:22:35 by rhunders          #+#    #+#             */
-/*   Updated: 2018/11/25 20:46:36 by rhunders         ###   ########.fr       */
+/*   Updated: 2018/11/26 00:33:31 by rhunders         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 long			va_arg_di(va_list ap, t_conv conv)
 {
 	if (!conv.modifier)
-		return ((long)va_arg(ap, int));
+		return ((long)((int)va_arg(ap, int)));
 	else if (conv.modifier == L)
 		return (va_arg(ap, long));
 	else if (conv.modifier == H)
@@ -29,7 +29,7 @@ long			va_arg_di(va_list ap, t_conv conv)
 unsigned long   va_arg_oux(va_list ap, t_conv conv)
 {
 	if (!conv.modifier)
-		return ((unsigned long)va_arg(ap, int));
+		return ((unsigned long)((unsigned int)va_arg(ap, int)));
 	else if (conv.modifier == L)
 		return (va_arg(ap, unsigned long));
 	else if (conv.modifier == H)
